refactor: rename the baseUI test to be reflective of mainAcitivity
