+++ code/app/src/androidTest/java/com/hamidat/nullpointersapp/AddMoodSocialSituationIntentTestMainActivity.java
@@ -3,8 +3,6 @@ package com.hamidat.nullpointersapp;
 import android.os.SystemClock;
 import android.util.Log;
 
-import androidx.test.espresso.contrib.RecyclerViewActions;
-import androidx.test.espresso.matcher.ViewMatchers;
 import androidx.test.filters.LargeTest;
 
 import com.hamidat.nullpointersapp.utils.testUtils.TestMoodHelper;
@@ -20,10 +18,10 @@ import static androidx.test.espresso.assertion.ViewAssertions.matches;
 import static androidx.test.espresso.matcher.ViewMatchers.isDisplayed;
 import static androidx.test.espresso.matcher.ViewMatchers.withId;
 import static androidx.test.espresso.matcher.ViewMatchers.withText;
-import static com.hamidat.nullpointersapp.SearchForOtherUsersIntentTest.withIndex;
+import static com.hamidat.nullpointersapp.SearchForOtherUsersIntentTestMainActivity.withIndex;
 
 @LargeTest
-public class AddMoodSocialSituationIntentTest extends BaseUITest {
+public class AddMoodSocialSituationIntentTestMainActivity extends BaseMainActivityUITest {
 
     private static final String TEST_REASON = "Testing social situation label.";
     private static final String EXPECTED_SOCIAL_SITUATION_LABEL = "Group";
