+++ code/app/src/androidTest/java/com/hamidat/nullpointersapp/ViewOtherUsersProfileIntentTestMainActivity.java
@@ -7,8 +7,6 @@ import static androidx.test.espresso.assertion.ViewAssertions.matches;
 import static androidx.test.espresso.matcher.ViewMatchers.isDisplayed;
 import static androidx.test.espresso.matcher.ViewMatchers.withId;
 import static androidx.test.espresso.matcher.ViewMatchers.withText;
-import static com.hamidat.nullpointersapp.SearchForOtherUsersIntentTest.withIndex;
-import static com.hamidat.nullpointersapp.androidTestHelpers.ViewActionsHelper.hasItemCount;
 import static org.hamcrest.Matchers.allOf;
 
 import android.os.SystemClock;
@@ -18,11 +16,11 @@ import androidx.test.espresso.contrib.RecyclerViewActions;
 
 import org.junit.Test;
 
-public class ViewThreeRecentMoodsIntentTest extends BaseUITest {
+public class ViewOtherUsersProfileIntentTestMainActivity extends BaseMainActivityUITest {
     private static final String TEST_SEARCH_QUERY = "hamihami";
 
     @Test
-    public void viewTheOtherUsersProfileForMoods() {
+    public void viewTheOtherUsersProfile() {
         // Click on search icon and ensure search edit text is displayed.
         onView(withId(R.id.ivSearch)).perform(click());
         onView(withId(R.id.etSearch)).check(matches(isDisplayed()));
@@ -41,12 +39,6 @@ public class ViewThreeRecentMoodsIntentTest extends BaseUITest {
         onView(allOf(withId(R.id.username_text), withText(TEST_SEARCH_QUERY)))
                 .check(matches(isDisplayed()));
 
-        // Wait for mood events to be fetched
-        SystemClock.sleep(3000);
-
-        // Assert that exactly 3 moods are loaded in rvMoodEvents
-        onView(withIndex(withId(R.id.rvMoodEvents), 1)).check(matches(hasItemCount(3)));
-
-        Log.d("ViewThreeRecentMoodsIntentTest", "3 mood events are visible");
+        Log.d("ViewOtherUsersProfileTest", "The other users profile is visible");
     }
 }
